Fix an issue with 32 bit transaction id's

+++ src/VPay.Ols.Processor/Models/TransactionIdLookup.cs
@@ -1,3 +1,3 @@
 ﻿namespace VPay.Ols.Processor.Models;
 
-public record TransactionIdLookup(int TransactionId);
+public record TransactionIdLookup(long TransactionId);
